+++ src/main/groovy/com/netgrif/application/engine/startup/ElasticsearchRunner.groovy
@@ -3,6 +3,7 @@ package com.netgrif.application.engine.startup
 import com.netgrif.application.engine.configuration.ElasticsearchConfiguration
 import com.netgrif.application.engine.configuration.properties.UriProperties
 import com.netgrif.application.engine.elastic.domain.ElasticCase
+import com.netgrif.application.engine.elastic.domain.ElasticPetriNet
 import com.netgrif.application.engine.elastic.domain.ElasticTask
 import com.netgrif.application.engine.elastic.service.interfaces.IElasticIndexService
 import com.netgrif.application.engine.petrinet.domain.UriNode
@@ -10,7 +11,6 @@ import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
-
 
 @Component
 @Slf4j
@@ -28,6 +28,9 @@ class ElasticsearchRunner extends AbstractOrderedCommandLineRunner {
     @Value('${spring.data.elasticsearch.port}')
     private int port
 
+    @Value('${spring.data.elasticsearch.index.petriNet}')
+    private String petriNetIndex
+
     @Value('${spring.data.elasticsearch.index.case}')
     private String caseIndex
 
@@ -42,8 +45,7 @@ class ElasticsearchRunner extends AbstractOrderedCommandLineRunner {
 
     @Autowired
     private ElasticsearchConfiguration properties
-
-
+    
     @Override
     void run(String... args) throws Exception {
         if (drop) {
@@ -53,6 +55,10 @@ class ElasticsearchRunner extends AbstractOrderedCommandLineRunner {
             }
             if (template.indexExists(taskIndex)) {
                 template.deleteIndex(ElasticTask.class)
+            }
+            if (!template.indexExists(petriNetIndex)) {
+                log.info "Creating Elasticsearch case index [${petriNetIndex}]"
+                template.createIndex(ElasticPetriNet.class)
             }
             try {
                 template.getAllDynamicIndexes().forEach(indexName -> {
@@ -84,19 +90,11 @@ class ElasticsearchRunner extends AbstractOrderedCommandLineRunner {
         } else {
             log.info "Elasticsearch task index exists [${taskIndex}]"
         }
-        if (!template.indexExists(uriProperties.index)) {
-            log.info "Creating Elasticsearch uri index [${uriProperties.index}]"
-            template.createIndex(UriNode.class)
-        } else {
-            log.info "Elasticsearch uri index exists [${uriProperties.index}]"
-        }
+        log.info("Updating Elasticsearch case mapping [${petriNetIndex}]")
+        template.putMapping(ElasticPetriNet.class)
         log.info("Updating Elasticsearch case mapping [${caseIndex}]")
         template.putMapping(ElasticCase.class)
         log.info("Updating Elasticsearch task mapping [${taskIndex}]")
         template.putMapping(ElasticTask.class)
-        log.info("Updating Elasticsearch uri mapping [${uriProperties.index}]")
-        template.putMapping(UriNode.class)
     }
-
-
 }