+++ src/main/java/com/netgrif/application/engine/elastic/service/interfaces/IElasticPetriNetMappingService.java
@@ -1,0 +1,8 @@
+package com.netgrif.application.engine.elastic.service.interfaces;
+
+import com.netgrif.application.engine.elastic.domain.ElasticPetriNet;
+import com.netgrif.application.engine.petrinet.domain.PetriNet;
+
+public interface IElasticPetriNetMappingService {
+    ElasticPetriNet transform(PetriNet net);
+}