+++ src/main/java/com/netgrif/application/engine/petrinet/domain/Transition.java
@@ -11,9 +11,11 @@ import com.netgrif.application.engine.petrinet.domain.layout.TaskLayout;
 import com.netgrif.application.engine.petrinet.domain.policies.AssignPolicy;
 import com.netgrif.application.engine.petrinet.domain.policies.DataFocusPolicy;
 import com.netgrif.application.engine.petrinet.domain.policies.FinishPolicy;
+import com.netgrif.application.engine.workflow.domain.triggers.AutoTrigger;
 import com.netgrif.application.engine.workflow.domain.triggers.Trigger;
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.data.annotation.Transient;
 import org.apache.lucene.analysis.CharArrayMap;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.Field;
@@ -89,6 +91,13 @@ public class Transition extends Node {
     @Setter
     private String defaultRoleId;
 
+    @Transient
+    private Boolean hasAutoTrigger;
+
+    @Getter
+    @Setter
+    private Map<String, String> tags;
+
     public Transition() {
         super();
         dataSet = new LinkedHashMap<>();
@@ -102,6 +111,7 @@ public class Transition extends Node {
         finishPolicy = FinishPolicy.MANUAL;
         events = new HashMap<>();
         assignedUserPolicy = new HashMap<>();
+        tags = new HashMap<>();
     }
 
     public void addDataSet(String field, Set<FieldBehavior> behavior, Map<DataEventType, DataEvent> events, FieldLayout layout, Component component){
@@ -242,6 +252,13 @@ public class Transition extends Node {
         events.put(event.getType(), event);
     }
 
+    public boolean hasAutoTrigger() {
+        if (hasAutoTrigger == null) {
+            hasAutoTrigger = this.getTriggers().stream().anyMatch(trigger -> trigger instanceof AutoTrigger);
+        }
+        return hasAutoTrigger;
+    }
+
     @Override
     public Transition clone() {
         Transition clone = new Transition();
@@ -263,6 +280,7 @@ public class Transition extends Node {
         clone.setFinishPolicy(finishPolicy);
         clone.setEvents(this.events == null ? null : events.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().clone())));
         clone.setAssignedUserPolicy(new HashMap<>(assignedUserPolicy));
+        clone.setTags(new HashMap<>(this.tags));
         clone.setDefaultRoleId(defaultRoleId);
         return clone;
     }