+++ src/main/java/com/netgrif/application/engine/workflow/domain/Task.java
@@ -23,13 +23,16 @@ import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.stereotype.Service;
 
+import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.util.*;
 
 @Document
 @AllArgsConstructor
 @Builder(builderMethodName = "with")
-public class Task {
+public class Task implements Serializable {
+
+    private static final long serialVersionUID = -7112277728921547546L;
 
     @Id
     @Builder.Default
@@ -188,6 +191,11 @@ public class Task {
     private Map<String, Boolean> assignedUserPolicy = new HashMap<>();
 
     private Map<String, Integer> consumedTokens = new HashMap<>();
+
+    @Getter
+    @Setter
+    @Builder.Default
+    private Map<String, String> tags = new HashMap<>();
 
     public Task() {
     }