+++ src/test/groovy/com/netgrif/application/engine/TestHelper.groovy
@@ -69,6 +69,8 @@ class TestHelper {
     private ImpersonationRunner impersonationRunner
 
     @Autowired
+    private UriRunner uriRunner
+    @Autowired
     private IPetriNetService petriNetService
 
     @Autowired