+++ CHANGELOG.md
@@ -5,7 +5,40 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-Full Changelog: [https://github.com/netgrif/application-engine/commits/v6.3.1](https://github.com/netgrif/application-engine/commits/v6.3.1)
+Full Changelog: [https://github.com/netgrif/application-engine/commits/v6.3.2](https://github.com/netgrif/application-engine/commits/v6.3.2)
+
+## [6.4.0](https://github.com/netgrif/application-engine/releases/tag/v6.4.0) (2024-01-23)
+
+### Fixed
+- [NAE-1908] NAE-1906 Improvements
+- [NAE-1937] Fix the problem with empty string in filter
+
+### Added
+- [NAE-1901] Taskref list rendering update
+- [NAE-1497] Frontend Actions
+- [NAE-1916] Serializable core objects
+- [NAE-1918] Tags on process resources
+- [NAE-1929] Data field type list of strings
+- [NAE-1936] Disable create case button using menu items
+
+## [6.3.3](https://github.com/netgrif/application-engine/releases/tag/v6.3.3) (2024-01-19)
+
+### Fixed
+
+- [NAE-1930] Modify Annotation in LdapUser Class for Indexing
+- [NAE-1931] Implementation of Index Re-Creation after MongoDB Database Drop
+- [NAE-1934] Remove devtools dependency
+- [NAE-1938] Slow Indexing of Elastic Tasks
+
+### Added
+
+- [NAE-1932] Enhancement of Redis Connector to Support Authentication
+
+## [6.3.2](https://github.com/netgrif/application-engine/releases/tag/v6.3.2) (2023-11-08)
+
+### Fixed
+
+- [NAE-1910] Event log is missing change in case's active places
 
 ## [6.3.1](https://github.com/netgrif/application-engine/releases/tag/v6.3.1) (2023-07-18)